+++ langsrv/handler-initialized.go
@@ -11,7 +11,7 @@ import (
 
 func initialized(context *glsp.Context, params *protocol.InitializedParams) error {
 	for _, root := range ls.workspaceRoots {
-		matches, err := world.Current.FS.Glob(path.Join(strings.TrimPrefix("file://", root), "*/*.lithia"))
+		matches, err := world.Current.FS.Glob(path.Join(strings.TrimPrefix("file://", root), "**/*.lithia"))
 		if err != nil {
 			ls.server.Log.Errorf("package detection failed, due %s", err)
 			continue